+++ django_migration_linter/management/utils.py
@@ -21,8 +21,10 @@ def register_linting_configuration_options(parser):
 
     parser.add_argument(
         "--warnings-as-errors",
-        action="store_true",
-        help="handle warnings as errors",
+        type=str,
+        nargs="*",
+        help="handle warnings as errors. Optionally specify the tests to handle as "
+        "errors (e.g. RUNPYTHON_REVERSIBLE)",
     )
 
 
@@ -35,3 +37,16 @@ def configure_logging(verbosity):
         logger.disabled = True
     else:
         logging.basicConfig(format="%(message)s")
+
+
+def extract_warnings_as_errors_option(warnings_as_errors):
+    if isinstance(warnings_as_errors, list):
+        warnings_as_errors_tests = warnings_as_errors
+        # If the option is specified but without any test codes,
+        # all warnings become errors
+        all_warnings_as_errors = len(warnings_as_errors) == 0
+    else:
+        warnings_as_errors_tests = None
+        all_warnings_as_errors = False
+
+    return warnings_as_errors_tests, all_warnings_as_errors